Refactor: integrate stream_reader and printer tests to reduce code duplication
